Move file listing logic into util file
